+++ Replication/CustomerIntelligence/Transforming/FirmStatisticsDto.cs
@@ -12,8 +12,8 @@ namespace NuClear.AdvancedSearch.Replication.CustomerIntelligence.Transforming
         {
             public long FirmId { get; set; }
             public long CategoryId { get; set; }
-            public int Hits { get; set; }
-            public int Shows { get; set; }
+            public long Hits { get; set; }
+            public long Shows { get; set; }
         }
     }
 }
